Change gateway service to use service traceprovider
This changes the gateway service to use the service traceprovider.
